Add workflow building and checking format of code
Workflow consists of two jobs. The `format` job checks if formatting of
the code follows conventions used in the organization. The `build` jobs
compiles the code.
Workflow will be executed for all PRs, merges to `main` and on schedule,
at midnight (UTC). It will also be possible to trigerr the flow
manually.