Add Countdown component
Since we are using countdown in so many places I decided to pull it out to a
component repo as a separate component instead using it as a hook. This
Countdown component doesn't really have JSX body - instead it takes a function
that returns the JSX Element as `children` prop. This way we can customize how
the component look and use needed values as hours, days etc in the function.