Deploy components to storybook
To help the design team participate in review of component pull
requests, we create a new job that builds the storybook based on the
`components` code and publishes the storybook to the GCP registry.
The job, when executed as a result of PR creation/update, will publish
the storybook to the `preview.storybook.threshold.network/<branch_name>`
bucket and will post a link to the page in a PR comment.
When job gets executed as a result of a push to `main` (PR merge), the
storybook will be uploaded to `storybook.threshold.network` bucket and
will be accessible by http://storybook.threshold.network/.