Fix `FilterTabs` component
We use the `selectedTabId` prop as a default value of `useState` hook.
In that case, if the `selectedTabId` prop changes the component state is
not updated because the `useState` hook is used only once not every time
the props change. We must call `setSelectedTabId` in `useEffect` hook
once the `selectedTabId` prop changes.